refactor: ajustando aos padroes
