+++ frontend/src/assets/images/rightArrow.png
