+++ frontend/src/assets/images/smallCat.png
