+++ frontend/src/components/back/index.tsx
@@ -1,12 +1,12 @@
 import { ConteinerImg, ImgVoltar, Volta, VoltarConteiner } from "./styles";
 
 
-export default function Voltar(){
+export default function Back(){
     return(
 
         <VoltarConteiner>
             <ConteinerImg>
-                <ImgVoltar source={require('../../assets/images/voltar.png')}/>
+                <ImgVoltar source={require('../../assets/images/backArrow.png')}/>
                 
             </ConteinerImg>
