+++ frontend/src/components/back/styles.tsx
