+++ frontend/src/components/blueConteiner/style.tsx
