+++ frontend/src/components/smallTitle/styles.tsx
