+++ frontend/src/components/title/styles.tsx
