+++ frontend/src/pages/screenLogo/styles.tsx
