+++ frontend/src/pages/screenRegisterBuyer/index.tsx
@@ -1,12 +1,14 @@
 import { Pressable } from "react-native";
-import BotaoLaranja from "../../components/botaoLaranja";
+
 import { Input } from "../../components/input";
-import TituloReduzido from "../../components/tituloReduzido";
+
 import { Patas } from "../login/styles";
 import { Check, InputsConteiner, Tela, Termos, TermosConteiner } from "./styles";
 import { useState } from "react";
+import SmallTitle from "../../components/smallTitle";
+import OrangeButton from "../../components/orangeButton";
 
-export default function CadastroComprador (){
+export default function ScreenRegisterBuyer (){
 
     const [trocar, setTrocar] = useState(true)
     
@@ -18,13 +20,13 @@ export default function CadastroComprador (){
 
     return(
         <Tela>
-        <TituloReduzido/>
+        <SmallTitle/>
         <InputsConteiner>
         <Input secureTextEntry={false} imagem={require('../../assets/images/userIcon.png')} placeholder="Nome ..."/>
         <Input secureTextEntry={false} imagem={require('../../assets/images/cpfIcon.png')} placeholder="CPF ..."/>
         <Input secureTextEntry={false} imagem={require('../../assets/images/arrobaIcon.png')} placeholder="E-mail ..."/>
-        <Input secureTextEntry={false} imagem={require('../../assets/images/cadeado.png')} placeholder="Senha ..."/>
-        <Input secureTextEntry={false} imagem={require('../../assets/images/cadeado.png')} placeholder="Confirmar Senha ..."/>
+        <Input secureTextEntry={false} imagem={require('../../assets/images/lock.png')} placeholder="Senha ..."/>
+        <Input secureTextEntry={false} imagem={require('../../assets/images/lock.png')} placeholder="Confirmar Senha ..."/>
         </InputsConteiner>
         <TermosConteiner>
             <Pressable onPress={Troca}>
@@ -32,7 +34,7 @@ export default function CadastroComprador (){
             </Pressable>
             <Termos>Concordo com os termos</Termos>
         </TermosConteiner>
-        <BotaoLaranja texto="Entrar"/>
+        <OrangeButton texto="Entrar"/>
         </Tela>
         
     )
