+++ frontend/src/pages/screenRegisterBuyer/styles.tsx
